% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_url.R
\name{build_gho_url}
\alias{build_gho_url}
\title{Construct GHO request URL}
\usage{
build_gho_url(
  base_url = "https://apps.who.int/gho/athena/api/",
  dimension = "GHO",
  code = NULL,
  filter = NULL,
  ...
)
}
\arguments{
\item{base_url}{Base URL of the API.}

\item{dimension}{Dimension of interest.}

\item{code}{A vector of requested codes.}

\item{filter}{A named list of filtering parameters.}

\item{...}{Other optional URL parameters.}
}
\value{
An URL as a string.
}
\description{
Construct GHO request URL
}
\keyword{internal}
