% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/display_attrs.R
\name{display_attribute_values}
\alias{display_attribute_values}
\alias{display_attributes}
\title{Display Attribute Informations}
\usage{
display_attribute_values(x, attribute)

display_attributes(x)
}
\arguments{
\item{x}{A GHO object.}

\item{attribute}{Name of an attribute.}
}
\value{
A character vector of the distinct values taken
  by the attribute.
}
\description{
Given a GHO object and an attribute name, return the
distinct values taken by the attribute.
}
