% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_gho}
\alias{filter_gho}
\title{Filter a GHO Object on Attribute Values}
\usage{
filter_gho(x, ...)
}
\arguments{
\item{x}{\code{gho} object to
filter.}

\item{...}{Logical predicates. Multiple conditions are
combined with &.}
}
\value{
A \code{gho} object.
}
\description{
Filter a GHO Object on Attribute Values
}
