% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attrs.R
\name{get_attrs}
\alias{get_attrs}
\title{Get Dimension Attributes}
\usage{
get_attrs(dimension = "GHO")
}
\arguments{
\item{dimension}{A GHO dimension.}
}
\value{
A \code{data_frame} of attributes.
}
\description{
Get Dimension Attributes
}

