% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_attrs.R
\name{get_attrs_}
\alias{get_attrs}
\alias{get_attrs_}
\title{Get Dimension Attributes}
\usage{
get_attrs_(xml_dim)

get_attrs(xml_dim)
}
\arguments{
\item{xml_dim}{A parsed XML list of \code{Code} elements.}
}
\value{
A \code{data_frame} of attributes.
}
\description{
Get Dimension Attributes
}

