% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gho.R
\name{get_gho}
\alias{get_gho}
\title{GET a GHO URL}
\usage{
get_gho(url = getOption("rgho.baseurl"))
}
\arguments{
\item{url}{the url to retrieve, given as a character
string.}
}
\value{
A \code{ODataQuery} object.
}
\description{
GET a GHO URL
}
\keyword{internal}
