% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codes.R
\name{get_gho_codes}
\alias{get_gho_codes}
\title{Returns GHO Codes for a Given Dimension}
\usage{
get_gho_codes(dimension = "GHO")
}
\arguments{
\item{dimension}{A GHO dimension.}

\item{x}{For \code{filter}, \code{gho} object to
filter.}

\item{...}{Logical predicates. Multiple conditions are
combined with &.}
}
\value{
GHO codes as a character vector, labels as a
  \code{label} attribute and other attributes in a \code{attrs}
  attribute, as a \code{data_frame}.
}
\description{
Returns GHO Codes for a Given Dimension
}

