% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_codes.R
\name{get_gho_codes_}
\alias{get_gho_codes}
\alias{get_gho_codes_}
\title{Returns GHO Codes for a Given Dimension}
\usage{
get_gho_codes_(dimension)

get_gho_codes(dimension)
}
\arguments{
\item{dimension}{A GHO dimension.}
}
\value{
GHO codes as a character vector, and labels as a \code{label} attribute.
}
\description{
Returns GHO Codes for a Given Dimension
}

