% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dimensions.R
\name{get_gho_dimensions}
\alias{get_gho_dimensions}
\title{Returns Available GHO Dimensions}
\usage{
get_gho_dimensions()
}
\value{
GHO dimensions as a character vector, and labels as a \code{label} attribute.
}
\description{
Returns Available GHO Dimensions
}
\examples{

get_gho_dimensions()

}

