% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{get_proxy}
\alias{get_proxy}
\title{Find Local Proxy Settings}
\usage{
get_proxy()
}
\value{
A proxy, or \code{NULL} if not proxy are found.
}
\description{
Tries to find the local proxy settings to access a given url.
}
\keyword{internal}
