% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{get_proxy_list}
\alias{get_proxy_list}
\title{Find Local Proxy Settings}
\usage{
get_proxy_list(url)
}
\arguments{
\item{url}{The url to access, as a character string.}
}
\value{
A list of proxy, or \code{NULL} if not proxy are found.
}
\description{
Tries to find the local proxy settings to access a given url.
}

