% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gho.R, R/man-gho.R, R/print.R
\name{get_gho_dimensions}
\alias{get_gho_dimensions}
\alias{get_gho_codes}
\alias{get_gho_values}
\alias{get_gho_indicators}
\alias{man-gho}
\alias{print.gho}
\title{Returns GHO Dimensions/Codes}
\usage{
get_gho_dimensions()

get_gho_codes(...)

get_gho_values(dimension = "GHO")

get_gho_indicators(filter = NULL)

\method{print}{gho}(x, n = options()$rgho.n, ...)
}
\arguments{
\item{...}{Extra arguments passed to print.}

\item{dimension}{A GHO dimension.}

\item{filter}{A named list of filtering parameters. Each parameter must be the
correct type.}

\item{x}{Object to print.}

\item{n}{Number of lines to print (if missing will print
the option \code{rgho.n} rows).}
}
\value{
GHO dimensions/values/indicators
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

This function was deprecated in favor of get_gho_values

Returns GHO dimensions, indicators, or, for a given dimension, values
}
\examples{

get_gho_dimensions()

get_gho_values(dimension = "GHO")

}
\keyword{internal}
