% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/proxy.R
\name{parse_proxy_string}
\alias{parse_proxy_string}
\title{Parse Proxy IP From a Character String}
\usage{
parse_proxy_string(proxy_string)
}
\arguments{
\item{proxy_string}{A list of proxy IP as a concatenated string.}
}
\value{
A list of proxy settings.
}
\description{
Given character string containig a list of proxy,
returns the list in a more computer readable form.
}
\keyword{internal}

