% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/values_attrs.R
\name{values_attrs}
\alias{values_attrs}
\title{Values Taken by an Attribute}
\usage{
values_attrs(x, attribute)
}
\arguments{
\item{x}{A GHO object.}

\item{attribute}{Name of an attribute.}
}
\value{
A character vector of the distinct values taken
  by the attribute.
}
\description{
Given a GHO object and an attribute name, return the
distinct values taken by the attribute.
}

