% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gho.R
\name{build_gho}
\alias{build_gho}
\title{Check and Build a gho Object}
\usage{
build_gho(x, labels, attrs = NULL)
}
\arguments{
\item{x}{A vector of codes.}

\item{labels}{A vector of labels.}

\item{attrs}{A \code{data_frame} of attributes.}
}
\value{
A \code{gho} object.
}
\description{
Check and Build a gho Object
}

