% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gho.R
\name{build_gho}
\alias{build_gho}
\title{Check and Build a gho Object}
\usage{
build_gho(x)
}
\arguments{
\item{x}{A ODataQuery object}
}
\value{
A \code{gho} object.
}
\description{
Check and Build a gho Object
}
\keyword{internal}
