% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gho.R
\name{get_gho_}
\alias{get_gho}
\alias{get_gho_}
\title{GET a GHO URL}
\usage{
get_gho_(url, verbose = FALSE, retry = 5)

get_gho(url, verbose = FALSE, retry = 5)
}
\arguments{
\item{url}{the url to retrieve, given as a character string.}

\item{verbose}{Verbose mode?}

\item{retry}{Maximum number of \code{GET} re-trials.}
}
\value{
The result from \code{\link[httr]{GET}}.
}
\description{
Given a url, tries to find local proxy settings and GET the content of the GHO page.
}
\details{
The function re-tries failed attempts using exponential backoff with jitter.
}
\keyword{internal}

