% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gho.R
\name{get_gho_}
\alias{get_gho}
\alias{get_gho_}
\title{GET a GHO URL}
\usage{
get_gho_(url)

get_gho(url)
}
\arguments{
\item{url}{the url to retrieve, given as a character string.}
}
\value{
The result from \code{httr} GET function.
}
\description{
Given a url, tries to find local proxy settings and GET the content of the GHO page.
}

