% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object.R
\name{get_gho_codes}
\alias{get_gho_codes}
\title{Returns GHO Codes for a Given Dimension}
\usage{
get_gho_codes(dimension = "GHO")
}
\arguments{
\item{dimension}{A GHO dimension.}
}
\value{
GHO codes as a character vector, labels as a
  \code{label} attribute and other attributes in a \code{attrs}
  attribute, as a \code{data_frame}.
}
\description{
Returns GHO Codes for a Given Dimension
}
\examples{

get_gho_codes(dimension = "GHO")

results <- get_gho_codes(dimension = "COUNTRY")
filter_gho(
  results,
  WHO_REGION_CODE == "EUR"
)

}

