% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R, R/get_codes.R
\name{filter_attrs}
\alias{filter_attrs}
\alias{get_gho_codes}
\title{Returns GHO Codes for a Given Dimension}
\usage{
filter_attrs(x, ...)

get_gho_codes(dimension = "GHO")
}
\arguments{
\item{x}{For \code{filter}, \code{gho} object to
filter.}

\item{...}{Logical predicates. Multiple conditions are
combined with &.}

\item{dimension}{A GHO dimension.}
}
\value{
GHO codes as a character vector, labels as a
  \code{label} attribute and other attributes in a \code{attrs}
  attribute, as a \code{data_frame}.
}
\description{
Returns GHO Codes for a Given Dimension
}
\examples{

get_gho_codes(dimension = "GHO")

results <- get_gho_codes(dimension = "COUNTRY")
filter_attrs(
  results,
  WHO_REGION_CODE == "EUR"
)

}

