% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R, R/get_codes.R
\name{filter_.gho}
\alias{filter_.gho}
\alias{get_gho_codes}
\alias{get_gho_codes_}
\title{Returns GHO Codes for a Given Dimension}
\usage{
\method{filter_}{gho}(.data, ..., .dots)

get_gho_codes_(dimension)

get_gho_codes(dimension)
}
\arguments{
\item{.data}{For \code{filter}, \code{gho} object to
filter.}

\item{...}{Logical predicates. Multiple conditions are
combined with &.}

\item{.dots}{Used to work around non-standard evaluation.
See vignette("nse", package = "dplyr") for details.}

\item{dimension}{A GHO dimension.}
}
\value{
GHO codes as a character vector, labels as a
  \code{label} attribute and other attributes in a \code{attrs}
  attribute, as a \code{data_frame}.
}
\description{
Returns GHO Codes for a Given Dimension
}

