% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_object.R, R/man-gho.R, R/print.R
\name{get_gho_dimensions}
\alias{get_gho_codes}
\alias{get_gho_dimensions}
\alias{man-gho}
\alias{print.gho}
\title{Returns GHO Dimensions/Codes}
\usage{
get_gho_dimensions()

get_gho_codes(dimension = "GHO")

\method{print}{gho}(x, n = options()$rgho.n, ...)
}
\arguments{
\item{dimension}{A GHO dimension.}

\item{x}{Object to print.}

\item{n}{Number of lines to print (if missing will print
the option \code{rgho.n} rows).}

\item{...}{Extra arguments passed to print.}
}
\value{
GHO dimensions/codes.
}
\description{
Returns GHO dimensions or, for a given dimension, codes.
}
\examples{

get_gho_dimensions()

get_gho_codes(dimension = "GHO")

}

