% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgho.R
\docType{package}
\name{rgho}
\alias{rgho}
\title{rgho: Access WHO Global Health Observatory Data From R}
\description{
Access WHO Global Health Observatory data from R via the
Athena web service, an application program interface
providing a simple query interface to the World Health
Organization's data and statistics content.
}
\section{Package Options}{


  \itemize{ \item \code{rgho.verbose} Should informations
  about the request be diplayed? Defaults to
  \code{FALSE}. \item \code{rgho.retry} How many times
  the request should be retried? Defaults to 5. \item
  \code{rgho.memotime} For how long should the results of
  a request be memoised, in seconds? Defaults to 3600. }
}

