% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rgho.R
\docType{package}
\name{rgho}
\alias{rgho}
\title{rgho: Access WHO Global Health Observatory Data From R}
\description{
Access WHO Global Health Observatory data from R via the
OData API, an application program interface
providing a simple query interface to the World Health
Organization's data and statistics content.
}
