% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{search_gho}
\alias{search_gho}
\alias{search_dimensions}
\alias{search_values}
\title{Search Codes or Dimensions Labels}
\usage{
search_gho(gho, x)

search_dimensions(x)

search_values(x, dimension = "GHO")
}
\arguments{
\item{gho}{An object from \code{\link{get_gho_dimensions}} or
\code{\link{get_gho_values}}.}

\item{x}{A character string to find in GHO labels.}

\item{dimension}{A GHO dimension where codes will be searched.}
}
\value{
A \code{GHO} object
}
\description{
Search Codes or Dimensions Labels
}
\examples{

search_values("neonatal", dimension = "GHO")

result <- get_gho_values(dimension = "REGION")
search_gho(result, "asia")

}
